requirements.txt
Flask==3.0.3
PyQt5==5.15.11
qrcode==7.4.2
pyngrok==7.2.0
twilio==9.2.4
pillow~=10.4.0
pytest~=8.3.3
pytest-qt~=4.4.0
